Backend_Django/proj_django/requirement.txt
asgiref==3.7.2
Django @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_cdse89o2w5/croots/recipe/django_1661292537238/work
django-cors-headers==4.3.1
djangorestframework==3.14.0
pipenv==2022.10.12
pytz==2023.3.post1
sqlparse @ file:///private/var/folders/k1/30mswbxs7r1g6zwn8y4fyt500000gp/T/abs_16ptvrhamb/croot/sqlparse_1690901943673/work
typing_extensions==4.9.0
virtualenv-clone==0.5.7
